refactor: update PBK model time resolution and plotting

+++ MCRA.Simulation/Calculators/KineticModelCalculation/PbpkModelCalculation/TargetExposurePerTimeUnit.cs
@@ -1,11 +1,11 @@
 ﻿namespace MCRA.Simulation.Calculators.KineticModelCalculation.PbpkModelCalculation {
     public sealed class TargetExposurePerTimeUnit {
-        public TargetExposurePerTimeUnit(int time = 0, double exposure = 0) {
+        public TargetExposurePerTimeUnit(double time = 0, double exposure = 0) {
             Time = time;
             Exposure = exposure;
         }
 
-        public int Time { get; set; }
+        public double Time { get; set; }
 
         public double Exposure { get; set; }
